add single file view from folder within sub app
