+++ package.json
@@ -62,6 +62,7 @@
     "wagmi": "^0.6.8"
   },
   "devDependencies": {
+    "@types/mime": "^3.0.1",
     "@types/mime-types": "^2.1.1",
     "autoprefixer": "^10.4.2",
     "eslint": "8.9.0",
